# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/productImages": {
            "get": {
                "tags": [
                    "ProductImages"
                ],
                "operationId": "078c3fe5fca5949b9cf5ac1374ae11af",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ProductImages Collection"
                    }
                }
            },
            "post": {
                "tags": [
                    "ProductImages"
                ],
                "summary": "Create productImage",
                "description": "Create productImage",
                "operationId": "1ff41921b7e3643256926f0acb80cd27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ProductImage Create"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/productImages/{id}": {
            "get": {
                "tags": [
                    "ProductImages"
                ],
                "description": "Show productImage info by id",
                "operationId": "331b0c76acf66b06dbc6008d531e2712",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ProductImage ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductImages"
                ],
                "operationId": "3ac26299a2bd1061c968e5a9705931be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ProductImage ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "ProductImage Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProductImages"
                ],
                "operationId": "8920583243656571ab440f7f59ad08b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ProductImage ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ProductImage Delete"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreProductImageRequest": {
                "title": "Store Glass request",
                "description": "Store Glass request body data",
                "properties": {
                    "product_id": {
                        "title": "product_id",
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "type": "string",
                        "example": "/images/products/test.jpg"
                    }
                },
                "type": "object"
            },
            "UpdateProductImageRequest": {
                "title": "Update Glass request",
                "description": "Update Glass request body data",
                "properties": {
                    "product_id": {
                        "title": "product_id",
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "type": "string",
                        "example": "/images/products/test.jpg"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}